+++ src/Responses/ResponsesStack.php
@@ -1,12 +1,13 @@
 <?php
 
-declare(strict_types=1);
+declare(strict_types = 1);
 
 namespace AvtoDev\JsonRpc\Responses;
 
-use Illuminate\Support\Collection;
+use ArrayIterator;
+use LogicException;
 
-class ResponsesStack extends Collection implements ResponsesStackInterface
+class ResponsesStack implements ResponsesStackInterface
 {
     /**
      * @var bool
@@ -14,32 +15,43 @@ class ResponsesStack extends Collection implements ResponsesStackInterface
     protected $is_batch;
 
     /**
-     * ResponsesStack constructor.
+     * @var array<ResponseInterface>
+     */
+    protected $responses = [];
+
+    /**
+     * Create a new ResponsesStack instance.
      *
      * @param bool                     $is_batch
      * @param array<ResponseInterface> $responses
      */
     public function __construct(bool $is_batch, array $responses = [])
     {
-        $this->is_batch = $is_batch;
+        $this->is_batch  = $is_batch;
+        $this->responses = $responses;
+    }
 
-        parent::__construct($responses);
+    /**
+     * @param bool                     $is_batch
+     * @param array<ResponseInterface> $responses
+     *
+     * @return self<ResponseInterface>
+     */
+    public static function make(bool $is_batch, array $responses = []): self
+    {
+        return new self($is_batch, $responses);
     }
 
     /**
      * Push response into stack.
      *
      * @param ResponseInterface $response
-     *
-     * @return self<ResponseInterface>
      */
-    public function push($response): self
+    public function push($response): void
     {
         if ($response instanceof ResponseInterface) {
-            $this->items[] = $response;
+            $this->responses[] = $response;
         }
-
-        return $this;
     }
 
     /**
@@ -48,5 +60,57 @@ class ResponsesStack extends Collection implements ResponsesStackInterface
     public function isBatch(): bool
     {
         return $this->is_batch;
+    }
+
+    /**
+     * {@inheritdoc}
+     */
+    public function all(): array
+    {
+        return $this->responses;
+    }
+
+    /**
+     * @return ArrayIterator<int, ResponseInterface>
+     */
+    public function getIterator(): ArrayIterator
+    {
+        return new ArrayIterator($this->responses);
+    }
+
+    /**
+     * {@inheritdoc}
+     */
+    public function count(): int
+    {
+        return count($this->responses);
+    }
+
+    /**
+     * {@inheritdoc}
+     */
+    public function isEmpty(): bool
+    {
+        return empty($this->responses);
+    }
+
+    /**
+     * {@inheritdoc}
+     */
+    public function isNotEmpty(): bool
+    {
+        return ! $this->isEmpty();
+    }
+
+    /**
+     * {@inheritdoc}
+     */
+    public function first(): ResponseInterface
+    {
+        if (\count($this->responses) === 0) {
+            throw new LogicException('Stack is empty');
+        }
+
+        return \reset($this->responses);
     }
 }