+++ src/ServiceProvider.php
@@ -12,13 +12,6 @@ use AvtoDev\JsonRpc\Factories\FactoryInterface;
 class ServiceProvider extends \Illuminate\Support\ServiceProvider
 {
     /**
-     * Indicates if loading of the provider is deferred.
-     *
-     * @var bool
-     */
-    protected $defer = false;
-
-    /**
      * Register RPC services.
      *
      * @return void